requirements.txt
speechrecognition
pyttsx3
pywhatkit
openai
pyaudio
python-dotenv
-e .

